model: move Amount and AccountType to new package
