+++ internal/filetransfer/correction_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/moov-io/paygate/internal"
 	"github.com/moov-io/paygate/internal/config"
 	"github.com/moov-io/paygate/internal/database"
+	"github.com/moov-io/paygate/internal/model"
 	"github.com/moov-io/paygate/internal/secrets"
 	"github.com/moov-io/paygate/pkg/id"
 
@@ -231,7 +232,7 @@ func TestController__handleNOCFile(t *testing.T) {
 		BankName:      "bank name",
 		Holder:        "holder",
 		HolderType:    internal.Individual,
-		Type:          internal.Checking,
+		Type:          model.Checking,
 		Status:        internal.DepositoryVerified,
 		Created:       base.NewTime(time.Now().Add(-1 * time.Second)),
 	}
