+++ internal/model/account.go
@@ -1,0 +1,43 @@
+// Copyright 2020 The Moov Authors
+// Use of this source code is governed by an Apache License
+// license that can be found in the LICENSE file.
+
+package model
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+)
+
+type AccountType string
+
+const (
+	Checking AccountType = "checking"
+	Savings  AccountType = "savings"
+)
+
+func (t AccountType) empty() bool {
+	return string(t) == ""
+}
+
+func (t AccountType) Validate() error {
+	switch t {
+	case Checking, Savings:
+		return nil
+	default:
+		return fmt.Errorf("AccountType(%s) is invalid", t)
+	}
+}
+
+func (t *AccountType) UnmarshalJSON(b []byte) error {
+	var s string
+	if err := json.Unmarshal(b, &s); err != nil {
+		return err
+	}
+	*t = AccountType(strings.ToLower(s))
+	if err := t.Validate(); err != nil {
+		return err
+	}
+	return nil
+}
