+++ internal/originators_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/moov-io/base"
 	"github.com/moov-io/paygate/internal/customers"
 	"github.com/moov-io/paygate/internal/database"
+	"github.com/moov-io/paygate/internal/model"
 	"github.com/moov-io/paygate/internal/secrets"
 	"github.com/moov-io/paygate/pkg/id"
 
@@ -164,7 +165,7 @@ func TestOriginators_CustomersError(t *testing.T) {
 		BankName:               "bank name",
 		Holder:                 "holder",
 		HolderType:             Individual,
-		Type:                   Checking,
+		Type:                   model.Checking,
 		RoutingNumber:          "123",
 		EncryptedAccountNumber: "151",
 		Status:                 DepositoryUnverified,
@@ -298,7 +299,7 @@ func TestOriginators__HTTPPost(t *testing.T) {
 	if err := depRepo.UpsertUserDepository(userID, &Depository{
 		ID:            id.Depository("foo"),
 		RoutingNumber: "987654320",
-		Type:          Checking,
+		Type:          model.Checking,
 		BankName:      "bank name",
 		Holder:        "holder",
 		HolderType:    Individual,
