+++ internal/transfers_tel_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/moov-io/base"
+	"github.com/moov-io/paygate/internal/model"
 	"github.com/moov-io/paygate/internal/secrets"
 	"github.com/moov-io/paygate/pkg/id"
 )
@@ -39,7 +40,7 @@ func TestTEL__createTELBatch(t *testing.T) {
 		BankName:      "foo bank",
 		Holder:        "jane doe",
 		HolderType:    Individual,
-		Type:          Checking,
+		Type:          model.Checking,
 		RoutingNumber: "121042882",
 		Status:        DepositoryVerified,
 		Metadata:      "jane doe checking",
@@ -58,7 +59,7 @@ func TestTEL__createTELBatch(t *testing.T) {
 		BankName:               "foo bank",
 		Holder:                 "john doe",
 		HolderType:             Individual,
-		Type:                   Savings,
+		Type:                   model.Savings,
 		RoutingNumber:          "231380104",
 		EncryptedAccountNumber: "2",
 		Status:                 DepositoryVerified,
@@ -72,7 +73,7 @@ func TestTEL__createTELBatch(t *testing.T) {
 		Identification:    "dddd",
 		Metadata:          "john doe",
 	}
-	amt, _ := NewAmount("USD", "100.00")
+	amt, _ := model.NewAmount("USD", "100.00")
 	transfer := &Transfer{
 		ID:                     TransferID(base.ID()),
 		Type:                   PushTransfer,
